+++ parserfn/AccessSitesLog.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace parserfn
+{
+    public class AccessedSitesLog
+    {
+        public string id { get; set; }
+        public string PartitionKey { get; set; }
+        public string Date { get; set; }
+        public string Url { get; set; }
+        public long MainModuleCount { get; set; }
+        public long SubModuleCount { get; set; }
+    }
+}
