+++ parserfn/JsonToDB.cs
@@ -11,15 +11,7 @@ using System.Collections.Generic;
 namespace parserfn
 {
     // Cosmos DB Entity
-    public class AccessedSitesLog
-    {
-        public string id { get; set; }
-        public string PartitionKey { get; set; }
-        public string Date { get; set; }
-        public string Url { get; set; }
-        public long MainModuleCount { get; set; }
-        public long SubModuleCount { get; set; }
-    }
+   
 
     public static class JsonToDB
     {
